+++ app/api/task_definitions_api.rb
@@ -6,6 +6,7 @@ class TaskDefinitionsApi < Grape::API
   helpers FileHelper
   helpers MimeCheckHelpers
   helpers Submission::GenerateHelpers
+  helpers FileStreamHelper
 
   before do
     authenticated?
@@ -551,8 +552,7 @@ class TaskDefinitionsApi < Grape::API
     end
 
     content_type 'application/pdf'
-    env['api.format'] = :binary
-    File.read(path)
+    stream_file path
   end
 
   desc 'Download the task resources'
@@ -579,8 +579,7 @@ class TaskDefinitionsApi < Grape::API
     end
     header['Access-Control-Expose-Headers'] = 'Content-Disposition'
 
-    env['api.format'] = :binary
-    File.read(path)
+    stream_file path
   end
 
   desc 'Download the task assessment resources'
@@ -607,7 +606,6 @@ class TaskDefinitionsApi < Grape::API
     end
     header['Access-Control-Expose-Headers'] = 'Content-Disposition'
 
-    env['api.format'] = :binary
-    File.read(path)
+    stream_file path
   end
 end